Organised User/Business components and minor changes
