--- src/pages/Product.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-const Product = () => {
-  return (
-    <>
-      Product
-    </>
-  )
-}
-
-export default Product
